Moved "Import Deck" feature from "collections" handling to "deck" handling
